Refactored Input into KeyboardInput and MouseInput
The linter was unhappy about the number of classes involved in Input,
I'm still not convinced this is a better result.
